added bottomtab
now need to start working on chat and individualchat
