+++ src/screens/ChatScreen.js
@@ -1,54 +1,66 @@
-import React, { useState, useCallback, useEffect } from "react";
-import { GiftedChat } from "react-native-gifted-chat";
-import db from "../../firebase";
-import { doc, onSnapshot, updateDoc, arrayUnion } from "firebase/firestore";
-import { useAuthentication } from "../utils/hooks/useAuthentication";
-import { View } from "react-native";
+import React, { useState, useEffect } from "react";
+import { Text, View, TouchableOpacity, StyleSheet } from "react-native";
+import { getAuth, signOut } from "firebase/auth";
 
+export default function ChatScreen({ navigation }) {
+  const [state, setState] = useState();
 
-export default function ChatScreen({ navigation, route }) {
-  const [messages, setMessages] = useState([]);
-  const chatRef = doc(db, "Chats", "myfirstchat");
+  const auth = getAuth();
+  const user = auth.currentUser;
 
-  const { user, userData } = useAuthentication();
+  console.log(user, "<--- user in the home screen");
 
-  useEffect(() => {
-    let unsubscribeFromNewSnapshots = onSnapshot(
-      chatRef,
-      (doc) => {
-        console.log("New Snapshot!");
-        setMessages(doc.data().messages);
-      }
+  if (user !== null) {
+    return (
+      <View style={styles.container}>
+        <TouchableOpacity
+          style={styles.logoutBtn}
+          onPress={() => {
+            signOut(auth)
+              .then(() => {
+                // Sign-out successful.
+                user = null;
+              })
+              .catch((error) => {
+                // An error happened.
+                // should we do something with that error??
+              });
+          }}
+        >
+          <Text style={styles.loginText}>sign out</Text>
+        </TouchableOpacity>
+
+        <Text>Hello, {user.email}! </Text>
+
+        <TouchableOpacity onPress={() => navigation.navigate("Chat")}>
+          <Text style={styles.item}>Chat</Text>
+        </TouchableOpacity>
+      </View>
     );
-
-    return function cleanupBeforeUnmounting() {
-      unsubscribeFromNewSnapshots();
-    };
-  }, []);
-
-  const onSend = useCallback((messages = []) => {
-    updateDoc(chatRef, {
-      // arrayUnion appends the message to the existing array
-      messages: arrayUnion(messages[0]),
-    });
-    // setMessages((previousMessages) =>
-    //   GiftedChat.append(previousMessages, messages)
-    // );
-  }, []);
-  if(!user || !userData) {
-    return <View></View>
   }
-  return (
-    <GiftedChat
-      messages={messages}
-      onSend={(messages) => onSend(messages)}
-      user={{
-        _id: user.uid,
-        name: userData.name
-      }}
-      inverted={false}
-      showUserAvatar={true}
-      renderUsernameOnMessage={true}
-    />
-  );
 }
+
+const styles = StyleSheet.create({
+  logoutBtn: {
+    width: "50%",
+    borderRadius: 25,
+    margin: 25,
+    height: 50,
+    alignItems: "center",
+    justifyContent: "center",
+    backgroundColor: "grey",
+    color: "white",
+  },
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+  },
+  item: {
+    padding: 10,
+    fontSize: 18,
+    height: 44,
+    backgroundColor: "yellow",
+    borderRadius: 25,
+    margin: 20,
+  },
+});
