+++ src/screens/SpotlightScreen.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { View, Text } from 'react-native'
+
+export default function SpotlightScreen() {
+  return (
+    <View>
+        <Text>
+            SpotlightScreen
+        </Text>
+    </View>
+  )
+}
