+++ src/screens/StoriesScreen.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { View, Text } from 'react-native'
+
+export default function StoriesScreen() {
+  return (
+    <View>
+        <Text>
+            StoriesScreen
+        </Text>
+    </View>
+  )
+}
